requirements.txt
python-arango
requests
beautifulsoup4
pandas
python-dotenv